Move activities and places into their own modules.
This enforces us not to close over variables from the outer scope when
creating activities, and also puts the url definitions in a single
location so they can be more self-contained.

Review: @joshmutus
